# Conflicts:
#	src/service/MockProjectWebApi/MockProjectWebApi.sln

+++ src/service/TRex/tests/netstandard/VSS.TRex.Tests/Cells/CellTests.cs
@@ -15,12 +15,12 @@ namespace VSS.TRex.Tests.Cells
         {
             Cell_NonStatic c = new Cell_NonStatic(0);
 
-            Assert.Equal((uint)0, c.PassCount);
+            Assert.Equal(0U, c.PassCount);
             Assert.True(c.IsEmpty, "Cell does not report itself as being empty of cell passes");
         }
 
         /// <summary>
-        /// Ensure the passcount allocation mechanism creates the appropriate number of entries
+        /// Ensure the pass count allocation mechanism creates the appropriate number of entries
         /// </summary>
         [Fact]
         public void Test_Cell_CellPassAllocation()
@@ -29,7 +29,7 @@ namespace VSS.TRex.Tests.Cells
 
             c.AllocatePasses(10);
 
-            Assert.Equal((uint)10, c.PassCount);
+            Assert.Equal(10U, c.PassCount);
         }
 
         /// <summary>
@@ -42,14 +42,14 @@ namespace VSS.TRex.Tests.Cells
 
             c.AddPass(CellPassTests.ATestCellPass());
 
-            Assert.Equal((uint)1, c.PassCount);
+            Assert.Equal(1U, c.PassCount);
 
             CellPass cp = CellPassTests.ATestCellPass();
             Assert.True(c.Passes[0].Equals(cp), "Added cell pass not the same content as it was constructed with");
 
             c.AddPass(CellPassTests.ATestCellPass2());
 
-            Assert.Equal((uint)2, c.PassCount);
+            Assert.Equal(2U, c.PassCount);
 
             CellPass cp2 = CellPassTests.ATestCellPass2();
             Assert.True(c.Passes[1].Equals(cp2), "Added cell pass not the same content as it was constructed with");
@@ -65,21 +65,21 @@ namespace VSS.TRex.Tests.Cells
 
             c.AddPass(CellPassTests.ATestCellPass2());
 
-            Assert.Equal((uint)1, c.PassCount);
+            Assert.Equal(1U, c.PassCount);
 
             CellPass cp2 = CellPassTests.ATestCellPass2();
             Assert.True(c.Passes[0].Equals(cp2), "Added cell pass not the same content as it was constructed with");
 
             c.AddPass(CellPassTests.ATestCellPass());
 
-            Assert.Equal((uint)2, c.PassCount);
+            Assert.Equal(2U, c.PassCount);
 
             CellPass cp1 = CellPassTests.ATestCellPass();
             Assert.True(c.Passes[0].Equals(cp1), "Added cell pass not the same content as it was constructed with");
         }
 
         /// <summary>
-        /// Ensure the topmostheight mechanism selects the correct 'top most' height in terms of 
+        /// Ensure the top most height mechanism selects the correct 'top most' height in terms of 
         /// the height from the most recently recorded cell pass in time
         /// </summary>
         [Fact]
@@ -91,7 +91,7 @@ namespace VSS.TRex.Tests.Cells
             c1.AddPass(CellPassTests.ATestCellPass());
             c1.AddPass(CellPassTests.ATestCellPass2());
 
-            Assert.Equal((uint)2, c1.PassCount);
+            Assert.Equal(2U, c1.PassCount);
             Assert.Equal(50, c1.TopMostHeight);
 
             // Add two cell passes in ascending order
@@ -100,7 +100,7 @@ namespace VSS.TRex.Tests.Cells
             c2.AddPass(CellPassTests.ATestCellPass2());
             c2.AddPass(CellPassTests.ATestCellPass());
 
-            Assert.Equal((uint)2, c2.PassCount);
+            Assert.Equal(2U, c2.PassCount);
             Assert.Equal(50, c2.TopMostHeight);
         }
 
@@ -172,10 +172,10 @@ namespace VSS.TRex.Tests.Cells
             c.AddPass(CellPassTests.ATestCellPass());
             c.AddPass(CellPassTests.ATestCellPass2());
 
-            Assert.Equal((uint)2, c.PassCount);
+            Assert.Equal(2U, c.PassCount);
 
             c.RemovePass(0);
-            Assert.Equal((uint)1, c.PassCount);
+            Assert.Equal(1U, c.PassCount);
 
             // Check the remaining pass is the one we think it is
             Assert.True(c.Passes[0].Equals(CellPassTests.ATestCellPass2()), "Remaining pass after removal is not the expected pass");
@@ -201,8 +201,8 @@ namespace VSS.TRex.Tests.Cells
             // Test integration of later cell pass to list with earlier cell pass resulting in 1 added and 0 modified
             c1.Integrate(c2.Passes, 0, c2.PassCount - 1, out uint addedCount, out uint modifiedCount);
 
-            Assert.Equal((uint)1, addedCount);
-            Assert.Equal((uint)0, modifiedCount);
+            Assert.Equal(1U, addedCount);
+            Assert.Equal(0U, modifiedCount);
 
             Assert.True(c2.PassCount == 1 && c2.Passes[0].Equals(cp2), "Integration unexpectedly modified source cell");
             Assert.True(c1.PassCount == 2 && c1.Passes[0].Equals(cp1) && c1.Passes[1].Equals(cp2),
@@ -227,8 +227,8 @@ namespace VSS.TRex.Tests.Cells
             // Test integration of the identical cell passes resulting in a single cell pass, 0 added and 0 modified (as identical cell passes are not a modification)
             c1.Integrate(c2.Passes, 0, c2.PassCount - 1, out uint addedCount, out uint modifiedCount);
 
-            Assert.Equal((uint)0, addedCount);
-            Assert.Equal((uint)0, modifiedCount);
+            Assert.Equal(0U, addedCount);
+            Assert.Equal(0U, modifiedCount);
 
             Assert.True(c2.PassCount == 1 && c2.Passes[0].Equals(cp1), "Integration unexpectedly modified source cell");
             Assert.True(c1.PassCount == 1 && c1.Passes[0].Equals(cp1),
@@ -242,10 +242,10 @@ namespace VSS.TRex.Tests.Cells
             // Test integration of the identical cell passes resulting in a single cell pass, 0 added and 1 modified (as identical cell passes are not a modification)
             c1.Integrate(c2.Passes, 0, c2.PassCount - 1, out addedCount, out modifiedCount);
 
-            Assert.Equal((uint)0, addedCount);
-            Assert.Equal((uint)1, modifiedCount);
+            Assert.Equal(0U, addedCount);
+            Assert.Equal(1U, modifiedCount);
 
-            Assert.Equal((uint)1, c1.PassCount);
+            Assert.Equal(1U, c1.PassCount);
         }
 
     }
